feat: major documentation improvements for API endpoints
- Rename 'API Endpoints' to 'Endpoint Reference' throughout docs
- Remove 'API' suffix from all endpoint documentation titles
- Remove duplicate table of contents, keep only Quick Reference
- Add tabbed code examples (Python, CLI, curl) for each endpoint
- Create endpoint_descriptions.json for maintaining endpoint descriptions
- Add descriptions for major endpoints and all resource tags
- Update index to show resource descriptions in a table
- Add sphinx-tabs extension for tabbed content support
- Improve documentation generator to use custom descriptions

The documentation now provides:
- Clear descriptions of what each resource/endpoint does
- Tabbed examples making it easy to switch between Python/CLI/curl
- Better organization without redundant navigation
- Comprehensive coverage of all 223+ API endpoints
