+++ docs/api/companysettings.rst
@@ -1,11 +1,5 @@
-CompanySettings API
-===================
-
-This section covers the 1 endpoints related to CompanySettings.
-
-.. contents::
-   :local:
-   :depth: 2
+CompanySettings
+===============
 
 Quick Reference
 ---------------
@@ -21,15 +15,11 @@ Quick Reference
      - /api/company/{companyId}/setupdata
      - 
 
-Endpoints
----------
 
 .. _get-apicompanycompanyidsetupdata:
 
 GET /api/company/{companyId}/setupdata
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-****
 
 **Parameters:**
 
@@ -39,21 +29,42 @@ GET /api/company/{companyId}/setupdata
 
 **Example Request:**
 
-Using curl:
+.. tabs::
 
-.. code-block:: bash
-   :linenos:
+   .. tab:: Python
 
-   curl -X GET \
-     -H 'Authorization: Bearer YOUR_API_TOKEN' \
-     'https://api.abconnect.co/api/company/ed282b80-54fe-4f42-bf1b-69103ce1f76c/setupdata'
+      .. code-block:: python
 
-Using AB CLI:
+         from ABConnect import ABConnectAPI
+         
+         # Initialize the API client
+         api = ABConnectAPI()
+         
+         # Make the API call
+         response = api.raw.get(
+             "/api/company/{companyId}/setupdata"
+         ,
+             companyId="ed282b80-54fe-4f42-bf1b-69103ce1f76c"
+         
+         )
+         
+         # Process the response
+         print(response)
 
-.. code-block:: bash
+   .. tab:: CLI
 
-   ab api raw get /api/company/{companyId}/setupdata \
-       companyId=ed282b80-54fe-4f42-bf1b-69103ce1f76c
+      .. code-block:: bash
+
+         ab api raw get /api/company/{companyId}/setupdata \
+             companyId=ed282b80-54fe-4f42-bf1b-69103ce1f76c
+
+   .. tab:: curl
+
+      .. code-block:: bash
+
+         curl -X GET \
+           -H 'Authorization: Bearer YOUR_API_TOKEN' \
+           'https://api.abconnect.co/api/company/ed282b80-54fe-4f42-bf1b-69103ce1f76c/setupdata'
 
 **Sample Response:**
 
@@ -64,9 +75,7 @@ Using AB CLI:
 
       {
         "status": "success",
-        "data": {
-          "message": "Operation completed successfully"
-        }
+        "data": {}
       }
 
 ----
