+++ docs/api/index.rst
@@ -1,11 +1,120 @@
-API Reference
-=============
+Endpoint Reference
+==================
 
-Complete API reference organized by resource type.
+Complete API endpoint reference organized by resource type.
+
+
+Available Resources
+-------------------
+
+.. list-table::
+   :header-rows: 1
+   :widths: 20 80
+
+   * - Resource
+     - Description
+   * - :doc:`account`
+     - 
+   * - :doc:`address`
+     - Manage and validate physical addresses used for shipping origins, destinations, and company locations.
+   * - :doc:`advancedsettings`
+     - 
+   * - :doc:`calendar`
+     - Access calendar events and scheduling information for jobs, deliveries, and appointments.
+   * - :doc:`carriererrormessage`
+     - 
+   * - :doc:`companies`
+     - Manage company records including customers, vendors, carriers, and franchisees. Companies are the core entities in the system that can place orders, provide services, or act as shipping partners.
+   * - :doc:`company_external_accounts`
+     - 
+   * - :doc:`companysettings`
+     - 
+   * - :doc:`contacthistory`
+     - 
+   * - :doc:`contactmerge`
+     - 
+   * - :doc:`contacts`
+     - Manage contact information for individuals associated with companies. Contacts can have various roles such as primary contact, billing contact, or operations contact.
+   * - :doc:`containerthicknessinches`
+     - 
+   * - :doc:`dashboard`
+     - Access dashboard widgets and analytics data for business intelligence and reporting.
+   * - :doc:`documents`
+     - Manage documents and files associated with jobs, companies, or contacts. Includes invoices, bills of lading, photos, and other supporting documentation.
+   * - :doc:`email`
+     - Send and manage email communications related to jobs, quotes, and customer notifications.
+   * - :doc:`globalsettings`
+     - 
+   * - :doc:`gridviews`
+     - 
+   * - :doc:`job`
+     - Manage shipping and logistics jobs. Jobs represent the movement of goods from origin to destination and include quotes, bookings, tracking, and completion.
+   * - :doc:`jobemail`
+     - 
+   * - :doc:`jobform`
+     - 
+   * - :doc:`jobfreightproviders`
+     - 
+   * - :doc:`jobintacct`
+     - 
+   * - :doc:`jobnote`
+     - 
+   * - :doc:`jobonhold`
+     - 
+   * - :doc:`jobparcelitems`
+     - 
+   * - :doc:`jobpayment`
+     - Process and track payments for jobs including credit cards, ACH, and other payment methods.
+   * - :doc:`jobrfq`
+     - 
+   * - :doc:`jobshipment`
+     - Manage shipment details within jobs including tracking, carrier assignments, and delivery confirmation.
+   * - :doc:`jobsign`
+     - 
+   * - :doc:`jobsms`
+     - 
+   * - :doc:`jobsmstemplate`
+     - 
+   * - :doc:`jobstatus`
+     - 
+   * - :doc:`jobtimeline`
+     - 
+   * - :doc:`jobtracking`
+     - Track real-time status and location of shipments throughout the delivery process.
+   * - :doc:`jobtrackingv2`
+     - 
+   * - :doc:`jobtrackingv3`
+     - 
+   * - :doc:`logbuffer`
+     - 
+   * - :doc:`lookup`
+     - Access system reference data and configuration values. Lookups provide standardized lists of values used throughout the system such as company types, job statuses, and service types.
+   * - :doc:`note`
+     - 
+   * - :doc:`notifications`
+     - Manage system notifications and alerts for users including email and SMS preferences.
+   * - :doc:`planner`
+     - 
+   * - :doc:`reports`
+     - Generate and access various business reports including financial summaries, operational metrics, and custom reports.
+   * - :doc:`rfq`
+     - 
+   * - :doc:`shipment`
+     - 
+   * - :doc:`stripewebhook`
+     - 
+   * - :doc:`truck`
+     - 
+   * - :doc:`twiliowebhook`
+     - 
+   * - :doc:`users`
+     - Manage user accounts and authentication. Users are individuals who can log into the system and perform various operations based on their permissions.
+   * - :doc:`values`
+     - 
 
 .. toctree::
    :maxdepth: 2
-   :caption: API Endpoints:
+   :hidden:
 
    account
    address
