+++ docs/api/jobtracking.rst
@@ -1,11 +1,7 @@
-JobTracking API
-===============
+JobTracking
+===========
 
-This section covers the 2 endpoints related to JobTracking.
-
-.. contents::
-   :local:
-   :depth: 2
+Track real-time status and location of shipments throughout the delivery process.
 
 Quick Reference
 ---------------
@@ -24,15 +20,11 @@ Quick Reference
      - /api/job/{jobDisplayId}/tracking/shipment/{proNumber}
      - 
 
-Endpoints
----------
 
 .. _get-apijobjobdisplayidtracking:
 
 GET /api/job/{jobDisplayId}/tracking
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-****
 
 **Parameters:**
 
@@ -42,21 +34,42 @@ GET /api/job/{jobDisplayId}/tracking
 
 **Example Request:**
 
-Using curl:
+.. tabs::
 
-.. code-block:: bash
-   :linenos:
+   .. tab:: Python
 
-   curl -X GET \
-     -H 'Authorization: Bearer YOUR_API_TOKEN' \
-     'https://api.abconnect.co/api/job/JOB-2024-001/tracking'
+      .. code-block:: python
 
-Using AB CLI:
+         from ABConnect import ABConnectAPI
+         
+         # Initialize the API client
+         api = ABConnectAPI()
+         
+         # Make the API call
+         response = api.raw.get(
+             "/api/job/{jobDisplayId}/tracking"
+         ,
+             jobDisplayId="2000000"
+         
+         )
+         
+         # Process the response
+         print(response)
 
-.. code-block:: bash
+   .. tab:: CLI
 
-   ab api raw get /api/job/{jobDisplayId}/tracking \
-       jobDisplayId=JOB-2024-001
+      .. code-block:: bash
+
+         ab api raw get /api/job/{jobDisplayId}/tracking \
+             jobDisplayId=2000000
+
+   .. tab:: curl
+
+      .. code-block:: bash
+
+         curl -X GET \
+           -H 'Authorization: Bearer YOUR_API_TOKEN' \
+           'https://api.abconnect.co/api/job/2000000/tracking'
 
 **Sample Response:**
 
@@ -67,9 +80,7 @@ Using AB CLI:
 
       {
         "status": "success",
-        "data": {
-          "message": "Operation completed successfully"
-        }
+        "data": {}
       }
 
 ----
@@ -78,8 +89,6 @@ Using AB CLI:
 
 GET /api/job/{jobDisplayId}/tracking/shipment/{proNumber}
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-****
 
 **Parameters:**
 
@@ -90,22 +99,45 @@ GET /api/job/{jobDisplayId}/tracking/shipment/{proNumber}
 
 **Example Request:**
 
-Using curl:
+.. tabs::
 
-.. code-block:: bash
-   :linenos:
+   .. tab:: Python
 
-   curl -X GET \
-     -H 'Authorization: Bearer YOUR_API_TOKEN' \
-     'https://api.abconnect.co/api/job/JOB-2024-001/tracking/shipment/example-value'
+      .. code-block:: python
 
-Using AB CLI:
+         from ABConnect import ABConnectAPI
+         
+         # Initialize the API client
+         api = ABConnectAPI()
+         
+         # Make the API call
+         response = api.raw.get(
+             "/api/job/{jobDisplayId}/tracking/shipment/{proNumber}"
+         ,
+             proNumber="example-value"
+         ,
+             jobDisplayId="2000000"
+         
+         )
+         
+         # Process the response
+         print(response)
 
-.. code-block:: bash
+   .. tab:: CLI
 
-   ab api raw get /api/job/{jobDisplayId}/tracking/shipment/{proNumber} \
-       proNumber=example-value \
-       jobDisplayId=JOB-2024-001
+      .. code-block:: bash
+
+         ab api raw get /api/job/{jobDisplayId}/tracking/shipment/{proNumber} \
+             proNumber=example-value \
+             jobDisplayId=2000000
+
+   .. tab:: curl
+
+      .. code-block:: bash
+
+         curl -X GET \
+           -H 'Authorization: Bearer YOUR_API_TOKEN' \
+           'https://api.abconnect.co/api/job/2000000/tracking/shipment/example-value'
 
 **Sample Response:**
 
@@ -116,9 +148,7 @@ Using AB CLI:
 
       {
         "status": "success",
-        "data": {
-          "message": "Operation completed successfully"
-        }
+        "data": {}
       }
 
 ----
