+++ docs/api/twiliowebhook.rst
@@ -1,11 +1,5 @@
-TwilioWebhook API
-=================
-
-This section covers the 1 endpoints related to TwilioWebhook.
-
-.. contents::
-   :local:
-   :depth: 2
+TwilioWebhook
+=============
 
 Quick Reference
 ---------------
@@ -21,36 +15,56 @@ Quick Reference
      - /api/webhooks/twilio/smsStatusCallback
      - 
 
-Endpoints
----------
 
 .. _post-apiwebhookstwiliosmsstatuscallback:
 
 POST /api/webhooks/twilio/smsStatusCallback
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-****
-
 **Example Request:**
 
-Using curl:
+.. tabs::
 
-.. code-block:: bash
-   :linenos:
+   .. tab:: Python
 
-   curl -X POST \
-     -H 'Authorization: Bearer YOUR_API_TOKEN' \
-     -H 'Content-Type: application/json' \
-     -d '{
-         "example": "data"
-     }' \
-     'https://api.abconnect.co/api/webhooks/twilio/smsStatusCallback'
+      .. code-block:: python
 
-Using AB CLI:
+         from ABConnect import ABConnectAPI
+         
+         # Initialize the API client
+         api = ABConnectAPI()
+         
+         # Make the API call
+         response = api.raw.post(
+             "/api/webhooks/twilio/smsStatusCallback"
+         ,
+             data=
+             {
+                 "example": "data"
+         }
+         
+         )
+         
+         # Process the response
+         print(response)
 
-.. code-block:: bash
+   .. tab:: CLI
 
-   ab api raw post /api/webhooks/twilio/smsStatusCallback
+      .. code-block:: bash
+
+         ab api raw post /api/webhooks/twilio/smsStatusCallback
+
+   .. tab:: curl
+
+      .. code-block:: bash
+
+         curl -X POST \
+           -H 'Authorization: Bearer YOUR_API_TOKEN' \
+           -H 'Content-Type: application/json' \
+           -d '{
+               "example": "data"
+           }' \
+           'https://api.abconnect.co/api/webhooks/twilio/smsStatusCallback'
 
 **Sample Response:**
 
@@ -62,11 +76,7 @@ Using AB CLI:
       {
         "id": "789e0123-e89b-12d3-a456-426614174002",
         "status": "created",
-        "message": "Resource created successfully",
-        "data": {
-          "id": "789e0123-e89b-12d3-a456-426614174002",
-          "created_at": "2024-01-20T10:00:00Z"
-        }
+        "message": "Resource created successfully"
       }
 
 ----
