+++ docs/api/values.rst
@@ -1,11 +1,5 @@
-Values API
-==========
-
-This section covers the 1 endpoints related to Values.
-
-.. contents::
-   :local:
-   :depth: 2
+Values
+======
 
 Quick Reference
 ---------------
@@ -21,15 +15,11 @@ Quick Reference
      - /api/Values
      - 
 
-Endpoints
----------
 
 .. _get-apivalues:
 
 GET /api/Values
 ~~~~~~~~~~~~~~~
-
-****
 
 **Parameters:**
 
@@ -39,20 +29,39 @@ GET /api/Values
 
 **Example Request:**
 
-Using curl:
+.. tabs::
 
-.. code-block:: bash
-   :linenos:
+   .. tab:: Python
 
-   curl -X GET \
-     -H 'Authorization: Bearer YOUR_API_TOKEN' \
-     'https://api.abconnect.co/api/Values'
+      .. code-block:: python
 
-Using AB CLI:
+         from ABConnect import ABConnectAPI
+         
+         # Initialize the API client
+         api = ABConnectAPI()
+         
+         # Make the API call
+         response = api.raw.get(
+             "/api/Values"
+         
+         )
+         
+         # Process the response
+         print(response)
 
-.. code-block:: bash
+   .. tab:: CLI
 
-   ab api raw get /api/Values
+      .. code-block:: bash
+
+         ab api raw get /api/Values
+
+   .. tab:: curl
+
+      .. code-block:: bash
+
+         curl -X GET \
+           -H 'Authorization: Bearer YOUR_API_TOKEN' \
+           'https://api.abconnect.co/api/Values'
 
 **Sample Response:**
