+++ docs/generate_api_docs.py
@@ -1,8 +1,8 @@
 #!/usr/bin/env python3
-"""Generate API documentation from swagger specification.
+"""Generate enhanced API documentation from swagger specification.
 
 This script creates comprehensive documentation for all API endpoints
-organized by swagger tags.
+organized by swagger tags with tabbed examples and descriptions.
 """
 
 import json
@@ -20,6 +20,14 @@ def sanitize_filename(name):
     return name.lower()
 
 
+def sanitize_anchor(text):
+    """Convert text to valid anchor name."""
+    # Remove special characters and convert to lowercase
+    text = re.sub(r'[^\w\s-]', '', text)
+    text = re.sub(r'[-\s]+', '-', text)
+    return text.lower()
+
+
 def format_parameter(param):
     """Format a parameter for documentation."""
     name = param.get('name', '')
@@ -33,181 +41,547 @@ def format_parameter(param):
     return f"- `{name}` ({param_type}, {param_in}){required_marker}: {description}"
 
 
-def format_bash_example(endpoint, base_url="https://api.abconnect.co"):
-    """Generate bash example for an endpoint."""
+def generate_realistic_example_value(param_name, param_type='string'):
+    """Generate realistic example values based on parameter name and type."""
+    param_lower = param_name.lower()
+    
+    # ID parameters
+    if 'id' in param_lower:
+        if 'display' in param_lower and 'job' in param_lower:
+            return '2000000'
+        elif 'job' in param_lower:
+            return 'JOB-2024-001'
+        elif 'company' in param_lower:
+            return 'ed282b80-54fe-4f42-bf1b-69103ce1f76c'
+        elif 'contact' in param_lower:
+            return '456e7890-e89b-12d3-a456-426614174001'
+        else:
+            return '789e0123-e89b-12d3-a456-426614174002'
+    
+    # Code parameters
+    elif 'code' in param_lower:
+        if 'company' in param_lower:
+            return 'TRAINING'
+        elif 'country' in param_lower:
+            return 'US'
+        else:
+            return 'CODE-001'
+    
+    # Name parameters
+    elif 'name' in param_lower:
+        if 'first' in param_lower:
+            return 'John'
+        elif 'last' in param_lower:
+            return 'Doe'
+        elif 'company' in param_lower:
+            return 'Acme Corporation'
+        else:
+            return 'Example Name'
+    
+    # Email parameters
+    elif 'email' in param_lower:
+        return 'user@example.com'
+    
+    # Phone parameters
+    elif 'phone' in param_lower:
+        return '+1-555-123-4567'
+    
+    # Date parameters
+    elif 'date' in param_lower or 'time' in param_lower:
+        if 'start' in param_lower:
+            return '2024-01-01T09:00:00Z'
+        elif 'end' in param_lower:
+            return '2024-01-31T17:00:00Z'
+        else:
+            return '2024-01-15T12:00:00Z'
+    
+    # Boolean parameters
+    elif param_type == 'boolean':
+        return 'true'
+    
+    # Number parameters
+    elif param_type in ['integer', 'number']:
+        if 'page' in param_lower:
+            return '1'
+        elif 'per_page' in param_lower or 'perpage' in param_lower:
+            return '20'
+        elif 'count' in param_lower:
+            return '10'
+        else:
+            return '100'
+    
+    # Default
+    else:
+        return 'example-value'
+
+
+def format_python_example(endpoint):
+    """Generate Python code example for an endpoint."""
     path = endpoint['path']
     method = endpoint['method']
     params = endpoint.get('parameters', [])
+    request_body = endpoint.get('requestBody', {})
     
-    # Replace path parameters with example values
-    example_path = path
+    # Build the code
+    code_lines = []
+    
+    # Import statement
+    code_lines.append("from ABConnect import ABConnectAPI")
+    code_lines.append("")
+    code_lines.append("# Initialize the API client")
+    code_lines.append("api = ABConnectAPI()")
+    code_lines.append("")
+    
+    # Prepare parameters
     path_params = []
     query_params = []
     
     for param in params:
         param_name = param.get('name', '')
         param_in = param.get('in', '')
+        schema = param.get('schema', {})
+        param_type = schema.get('type', 'string')
         
         if param_in == 'path':
-            # Use example values for common parameters
-            if 'id' in param_name.lower():
-                example_value = '123e4567-e89b-12d3-a456-426614174000'
-            elif 'displayid' in param_name.lower():
-                example_value = 'JOB-2024-001'
-            elif 'code' in param_name.lower():
-                example_value = 'ABC123'
-            else:
-                example_value = 'example-value'
-            
-            example_path = example_path.replace(f'{{{param_name}}}', example_value)
+            example_value = generate_realistic_example_value(param_name, param_type)
             path_params.append((param_name, example_value))
-            
-        elif param_in == 'query':
-            # Add common query parameters
-            if param_name == 'page':
-                query_params.append(f'{param_name}=1')
-            elif param_name == 'per_page' or param_name == 'perPage':
-                query_params.append(f'{param_name}=20')
-            elif param.get('required', False):
-                query_params.append(f'{param_name}=value')
+        elif param_in == 'query' and (param.get('required', False) or param_name in ['page', 'per_page', 'perPage']):
+            example_value = generate_realistic_example_value(param_name, param_type)
+            query_params.append((param_name, example_value))
+    
+    # Build the API call
+    code_lines.append("# Make the API call")
+    code_lines.append(f"response = api.raw.{method.lower()}(")
+    code_lines.append(f'    "{path}"')
+    
+    # Add path parameters
+    for param_name, value in path_params:
+        if param_type == 'string':
+            code_lines.append(f',\n    {param_name}="{value}"')
+        else:
+            code_lines.append(f',\n    {param_name}={value}')
+    
+    # Add query parameters
+    for param_name, value in query_params:
+        if param_type == 'string':
+            code_lines.append(f',\n    {param_name}="{value}"')
+        else:
+            code_lines.append(f',\n    {param_name}={value}')
+    
+    # Add request body for POST/PUT/PATCH
+    if method in ['POST', 'PUT', 'PATCH']:
+        content = request_body.get('content', {})
+        if 'application/json' in content:
+            schema = content['application/json'].get('schema', {})
+            sample_body = generate_sample_request_body(schema)
+            code_lines.append(',\n    data=')
+            body_lines = json.dumps(sample_body, indent=8).split('\n')
+            for i, line in enumerate(body_lines):
+                if i == 0:
+                    code_lines.append(f'    {line}')
+                else:
+                    code_lines.append(line)
+    
+    code_lines.append("\n)")
+    code_lines.append("")
+    code_lines.append("# Process the response")
+    code_lines.append("print(response)")
+    
+    return '\n'.join(code_lines)
+
+
+def format_cli_example(endpoint):
+    """Generate AB CLI example for an endpoint."""
+    path = endpoint['path']
+    method = endpoint['method']
+    params = endpoint.get('parameters', [])
+    
+    # Build the command
+    cmd_parts = [f"ab api raw {method.lower()} {path}"]
+    
+    path_params = []
+    query_params = []
+    
+    for param in params:
+        param_name = param.get('name', '')
+        param_in = param.get('in', '')
+        schema = param.get('schema', {})
+        param_type = schema.get('type', 'string')
+        
+        if param_in == 'path':
+            example_value = generate_realistic_example_value(param_name, param_type)
+            path_params.append(f"{param_name}={example_value}")
+        elif param_in == 'query' and (param.get('required', False) or param_name in ['page', 'per_page', 'perPage']):
+            example_value = generate_realistic_example_value(param_name, param_type)
+            query_params.append(f"{param_name}={example_value}")
+    
+    # Add parameters
+    all_params = path_params + query_params
+    if all_params:
+        cmd_parts.extend([f"    {param}" for param in all_params])
+    
+    return ' \\\n'.join(cmd_parts)
+
+
+def format_curl_example(endpoint, base_url="https://api.abconnect.co"):
+    """Generate curl example for an endpoint."""
+    path = endpoint['path']
+    method = endpoint['method']
+    params = endpoint.get('parameters', [])
+    request_body = endpoint.get('requestBody', {})
+    
+    # Replace path parameters with example values
+    example_path = path
+    query_params = []
+    
+    for param in params:
+        param_name = param.get('name', '')
+        param_in = param.get('in', '')
+        schema = param.get('schema', {})
+        param_type = schema.get('type', 'string')
+        
+        if param_in == 'path':
+            example_value = generate_realistic_example_value(param_name, param_type)
+            example_path = example_path.replace(f'{{{param_name}}}', example_value)
+        elif param_in == 'query' and (param.get('required', False) or param_name in ['page', 'per_page', 'perPage']):
+            example_value = generate_realistic_example_value(param_name, param_type)
+            query_params.append(f'{param_name}={example_value}')
     
     # Build the curl command
-    curl_cmd = f"curl -X {method}"
-    
-    # Add headers
-    curl_cmd += " \\\n  -H 'Authorization: Bearer YOUR_TOKEN'"
+    curl_lines = [f"curl -X {method} \\"]
+    curl_lines.append("  -H 'Authorization: Bearer YOUR_API_TOKEN' \\")
     
     if method in ['POST', 'PUT', 'PATCH']:
-        curl_cmd += " \\\n  -H 'Content-Type: application/json'"
-        curl_cmd += " \\\n  -d '{"
-        curl_cmd += "\n    \"key\": \"value\""
-        curl_cmd += "\n  }'"
+        curl_lines.append("  -H 'Content-Type: application/json' \\")
+        
+        # Generate sample request body
+        content = request_body.get('content', {})
+        if 'application/json' in content:
+            schema = content['application/json'].get('schema', {})
+            sample_body = generate_sample_request_body(schema)
+            curl_lines.append("  -d '{")
+            body_lines = json.dumps(sample_body, indent=2).split('\n')
+            for i, line in enumerate(body_lines[1:-1]):  # Skip first { and last }
+                if i == len(body_lines) - 3:  # Last line
+                    curl_lines.append(f"    {line}")
+                else:
+                    curl_lines.append(f"    {line}")
+            curl_lines.append("  }' \\")
     
     # Add URL with query parameters
     url = f"{base_url}{example_path}"
     if query_params:
         url += "?" + "&".join(query_params)
     
-    curl_cmd += f" \\\n  '{url}'"
+    curl_lines.append(f"  '{url}'")
     
-    # Also add ab CLI example
-    ab_cmd = f"ab api raw {method.lower()} {path}"
-    if path_params:
-        for name, value in path_params:
-            ab_cmd += f" {name}={value}"
-    if query_params:
-        ab_cmd += " " + " ".join(query_params)
+    return '\n'.join(curl_lines)
+
+
+def generate_sample_request_body(schema):
+    """Generate a realistic sample request body based on schema."""
+    if not schema:
+        return {"example": "data"}
     
-    return curl_cmd, ab_cmd
+    # Common request patterns
+    properties = schema.get('properties', {})
+    
+    sample = {}
+    for prop_name, prop_schema in properties.items():
+        prop_type = prop_schema.get('type', 'string')
+        sample[prop_name] = generate_realistic_example_value(prop_name, prop_type)
+    
+    return sample if sample else {"example": "data"}
+
+
+def load_response_examples():
+    """Load response examples from the response_examples.json file."""
+    response_examples_path = Path(__file__).parent / "api" / "response_examples.json"
+    if response_examples_path.exists():
+        with open(response_examples_path, 'r') as f:
+            return json.load(f)
+    return {}
+
+
+def load_endpoint_descriptions():
+    """Load endpoint descriptions from the endpoint_descriptions.json file."""
+    descriptions_path = Path(__file__).parent / "api" / "endpoint_descriptions.json"
+    if descriptions_path.exists():
+        with open(descriptions_path, 'r') as f:
+            return json.load(f)
+    return {"endpoints": {}, "tags": {}}
+
+
+# Global variables
+RESPONSE_EXAMPLES = None
+ENDPOINT_DESCRIPTIONS = None
+
+
+def get_endpoint_description(path, method):
+    """Get description for a specific endpoint."""
+    global ENDPOINT_DESCRIPTIONS
+    if ENDPOINT_DESCRIPTIONS is None:
+        ENDPOINT_DESCRIPTIONS = load_endpoint_descriptions()
+    
+    endpoints = ENDPOINT_DESCRIPTIONS.get('endpoints', {})
+    endpoint_data = endpoints.get(path, {}).get(method.lower(), {})
+    
+    return endpoint_data.get('summary', ''), endpoint_data.get('description', '')
+
+
+def get_tag_description(tag):
+    """Get description for a tag."""
+    global ENDPOINT_DESCRIPTIONS
+    if ENDPOINT_DESCRIPTIONS is None:
+        ENDPOINT_DESCRIPTIONS = load_endpoint_descriptions()
+    
+    tags = ENDPOINT_DESCRIPTIONS.get('tags', {})
+    return tags.get(tag, '')
+
+
+def get_endpoint_key(path, method):
+    """Generate a key for looking up response examples."""
+    # Normalize path for lookup
+    # Remove path parameters
+    normalized_path = re.sub(r'\{[^}]+\}', '{}', path)
+    
+    # Common endpoint patterns
+    if '/companies/{}/details' in normalized_path:
+        return 'companies.get_company_details'
+    elif '/companies/{}/fulldetails' in normalized_path and method == 'GET':
+        return 'companies.get_company_fulldetails'
+    elif '/companies/{}' in normalized_path and method == 'GET':
+        return 'companies.get_company_by_id'
+    elif '/companies/search' in normalized_path:
+        return 'companies.get_companies_search'
+    elif '/companies/availableByCurrentUser' in normalized_path:
+        return 'companies.get_companies_availableByCurrentUser'
+    
+    # Default key based on path
+    tag = path.split('/')[2] if len(path.split('/')) > 2 else 'generic'
+    operation = path.split('/')[-1].replace('{', '').replace('}', '')
+    return f"{tag}.{method.lower()}_{operation}"
 
 
 def generate_sample_response(endpoint):
-    """Generate a sample response based on endpoint details."""
-    # This is a simplified version - in reality you'd want to use the schema
+    """Generate a sample response using actual API responses when available."""
+    global RESPONSE_EXAMPLES
+    if RESPONSE_EXAMPLES is None:
+        RESPONSE_EXAMPLES = load_response_examples()
+    
+    path = endpoint['path']
+    method = endpoint['method']
+    
+    # Try to find a real response example
+    endpoint_key = get_endpoint_key(path, method)
+    
+    # Check for exact match first
+    for tag, endpoints in RESPONSE_EXAMPLES.items():
+        if tag == 'generic':
+            continue
+        for key, example in endpoints.items():
+            if example.get('endpoint') == path and example.get('method') == method:
+                response = example.get('response', {})
+                return json.dumps(response, indent=2)
+    
+    # Check for pattern match
+    parts = endpoint_key.split('.')
+    if len(parts) == 2:
+        tag, operation = parts
+        if tag in RESPONSE_EXAMPLES and operation in RESPONSE_EXAMPLES[tag]:
+            response = RESPONSE_EXAMPLES[tag][operation].get('response', {})
+            return json.dumps(response, indent=2)
+    
+    # Fallback to generic responses based on endpoint
     responses = endpoint.get('responses', {})
     
-    if '200' in responses:
-        response = responses['200']
-        content = response.get('content', {})
-        
-        if 'application/json' in content:
-            schema = content['application/json'].get('schema', {})
-            
-            # Generate sample based on common patterns
-            if 'array' in str(schema):
-                return json.dumps([
-                    {
-                        "id": "123e4567-e89b-12d3-a456-426614174000",
-                        "name": "Example Item",
-                        "created": "2024-01-01T00:00:00Z"
-                    }
-                ], indent=2)
-            else:
-                return json.dumps({
-                    "id": "123e4567-e89b-12d3-a456-426614174000",
-                    "status": "success",
-                    "data": {
-                        "example": "response"
-                    }
-                }, indent=2)
+    # Common response patterns based on endpoint
+    if 'search' in path.lower() or (method == 'GET' and path.endswith('s')):
+        # List response
+        if 'generic' in RESPONSE_EXAMPLES and 'empty_list' in RESPONSE_EXAMPLES['generic']:
+            return json.dumps(RESPONSE_EXAMPLES['generic']['empty_list'], indent=2)
+        return json.dumps([], indent=2)
     
-    return json.dumps({"message": "Success"}, indent=2)
+    elif '{id}' in path or 'details' in path:
+        # Single item response
+        if 'generic' in RESPONSE_EXAMPLES and 'empty_object' in RESPONSE_EXAMPLES['generic']:
+            return json.dumps(RESPONSE_EXAMPLES['generic']['empty_object'], indent=2)
+        return json.dumps({}, indent=2)
+    
+    elif method == 'POST':
+        # Create response
+        return json.dumps({
+            "id": "789e0123-e89b-12d3-a456-426614174002",
+            "status": "created",
+            "message": "Resource created successfully"
+        }, indent=2)
+    
+    elif method in ['PUT', 'PATCH']:
+        # Update response
+        return json.dumps({
+            "id": "123e4567-e89b-12d3-a456-426614174000",
+            "status": "updated",
+            "message": "Resource updated successfully"
+        }, indent=2)
+    
+    elif method == 'DELETE':
+        # Delete response
+        return json.dumps({
+            "status": "success",
+            "message": "Resource deleted successfully"
+        }, indent=2)
+    
+    else:
+        # Generic success response
+        return json.dumps({
+            "status": "success",
+            "data": {}
+        }, indent=2)
 
 
 def generate_tag_documentation(tag_name, endpoints, tag_description=""):
-    """Generate RST documentation for a tag."""
+    """Generate enhanced RST documentation for a tag."""
     doc = []
     
-    # Title
-    title = f"{tag_name} API"
+    # Title (without 'API' suffix)
+    title = tag_name
     doc.append(title)
     doc.append("=" * len(title))
     doc.append("")
     
-    # Description
-    if tag_description:
+    # Tag description
+    custom_description = get_tag_description(tag_name)
+    if custom_description:
+        doc.append(custom_description)
+        doc.append("")
+    elif tag_description:
         doc.append(tag_description)
         doc.append("")
     
-    doc.append(f"This section covers the {len(endpoints)} endpoints related to {tag_name}.")
+    # Quick reference table
+    doc.append("Quick Reference")
+    doc.append("---------------")
     doc.append("")
+    doc.append(".. list-table::")
+    doc.append("   :header-rows: 1")
+    doc.append("   :widths: 10 40 50")
+    doc.append("")
+    doc.append("   * - Method")
+    doc.append("     - Endpoint")
+    doc.append("     - Description")
     
-    # Table of contents
-    doc.append(".. contents::")
-    doc.append("   :local:")
-    doc.append("   :depth: 2")
+    for endpoint in endpoints:
+        method = endpoint['method']
+        path = endpoint['path']
+        
+        # Get custom summary or use swagger summary
+        custom_summary, _ = get_endpoint_description(path, method)
+        summary = custom_summary or endpoint.get('summary', '')
+        
+        doc.append(f"   * - {method}")
+        doc.append(f"     - {path}")
+        doc.append(f"     - {summary}")
+    
+    doc.append("")
     doc.append("")
     
     # Process each endpoint
     for endpoint in endpoints:
         operation_id = endpoint.get('operationId', '')
-        summary = endpoint.get('summary', 'No summary available')
-        description = endpoint.get('description', '')
         path = endpoint['path']
         method = endpoint['method']
         parameters = endpoint.get('parameters', [])
         
-        # Endpoint title
+        # Get custom descriptions
+        custom_summary, custom_description = get_endpoint_description(path, method)
+        summary = custom_summary or endpoint.get('summary', '')
+        description = custom_description or endpoint.get('description', '')
+        
+        # Create anchor ID for cross-referencing
+        anchor_id = sanitize_anchor(f"{method}-{path}")
+        
+        # Endpoint title with anchor
+        doc.append(f".. _{anchor_id}:")
+        doc.append("")
         endpoint_title = f"{method} {path}"
         doc.append(endpoint_title)
-        doc.append("-" * len(endpoint_title))
+        doc.append("~" * len(endpoint_title))
         doc.append("")
         
         # Summary and description
-        doc.append(summary)
-        if description and description != summary:
+        if summary:
+            doc.append(f"**{summary}**")
             doc.append("")
+        
+        if description:
             doc.append(description)
-        doc.append("")
+            doc.append("")
         
         # Parameters
         if parameters:
             doc.append("**Parameters:**")
             doc.append("")
-            for param in parameters:
-                doc.append(format_parameter(param))
-            doc.append("")
+            
+            # Group parameters by type
+            path_params = [p for p in parameters if p.get('in') == 'path']
+            query_params = [p for p in parameters if p.get('in') == 'query']
+            header_params = [p for p in parameters if p.get('in') == 'header']
+            
+            if path_params:
+                doc.append("*Path Parameters:*")
+                doc.append("")
+                for param in path_params:
+                    doc.append(format_parameter(param))
+                doc.append("")
+            
+            if query_params:
+                doc.append("*Query Parameters:*")
+                doc.append("")
+                for param in query_params:
+                    doc.append(format_parameter(param))
+                doc.append("")
+            
+            if header_params:
+                doc.append("*Header Parameters:*")
+                doc.append("")
+                for param in header_params:
+                    doc.append(format_parameter(param))
+                doc.append("")
         
-        # Request example
+        # Example Request with tabs
         doc.append("**Example Request:**")
         doc.append("")
-        
-        # Generate bash examples
-        curl_example, ab_example = format_bash_example(endpoint)
-        
-        # Curl example with copy button
-        doc.append(".. code-block:: bash")
-        doc.append("   :caption: Using curl")
+        doc.append(".. tabs::")
         doc.append("")
+        
+        # Python tab
+        doc.append("   .. tab:: Python")
+        doc.append("")
+        doc.append("      .. code-block:: python")
+        doc.append("")
+        python_example = format_python_example(endpoint)
+        for line in python_example.split('\n'):
+            doc.append(f"         {line}")
+        doc.append("")
+        
+        # CLI tab
+        doc.append("   .. tab:: CLI")
+        doc.append("")
+        doc.append("      .. code-block:: bash")
+        doc.append("")
+        cli_example = format_cli_example(endpoint)
+        for line in cli_example.split('\n'):
+            doc.append(f"         {line}")
+        doc.append("")
+        
+        # Curl tab
+        doc.append("   .. tab:: curl")
+        doc.append("")
+        doc.append("      .. code-block:: bash")
+        doc.append("")
+        curl_example = format_curl_example(endpoint)
         for line in curl_example.split('\n'):
-            doc.append(f"   {line}")
-        doc.append("")
-        
-        # AB CLI example
-        doc.append(".. code-block:: bash")
-        doc.append("   :caption: Using AB CLI")
-        doc.append("")
-        doc.append(f"   {ab_example}")
+            doc.append(f"         {line}")
         doc.append("")
         
         # Sample response (collapsible)
@@ -216,19 +590,23 @@ def generate_tag_documentation(tag_name, endpoints, tag_description=""):
         doc.append(".. toggle::")
         doc.append("")
         doc.append("   .. code-block:: json")
+        doc.append("      :linenos:")
         doc.append("")
         
         sample_response = generate_sample_response(endpoint)
         for line in sample_response.split('\n'):
             doc.append(f"      {line}")
         doc.append("")
+        
+        # Add a separator between endpoints
+        doc.append("----")
         doc.append("")
     
     return "\n".join(doc)
 
 
 def main():
-    """Generate API documentation."""
+    """Generate enhanced API documentation."""
     # Load swagger specification
     swagger_path = Path(__file__).parent.parent / "ABConnect" / "base" / "swagger.json"
     with open(swagger_path, 'r') as f:
@@ -242,7 +620,7 @@ def main():
     tag_groups = defaultdict(list)
     tag_descriptions = {}
     
-    # Get tag descriptions
+    # Get tag descriptions from swagger
     if 'tags' in swagger:
         for tag in swagger['tags']:
             tag_descriptions[tag['name']] = tag.get('description', '')
@@ -268,14 +646,36 @@ def main():
     
     # Generate index file
     index_content = []
-    index_content.append("API Reference")
-    index_content.append("=============")
+    index_content.append("Endpoint Reference")
+    index_content.append("==================")
     index_content.append("")
-    index_content.append("Complete API reference organized by resource type.")
+    index_content.append("Complete API endpoint reference organized by resource type.")
+    index_content.append("")
+    index_content.append("")
+    index_content.append("Available Resources")
+    index_content.append("-------------------")
+    index_content.append("")
+    
+    # Add a table of resources with descriptions
+    descriptions = load_endpoint_descriptions()
+    tag_descs = descriptions.get('tags', {})
+    
+    index_content.append(".. list-table::")
+    index_content.append("   :header-rows: 1")
+    index_content.append("   :widths: 20 80")
+    index_content.append("")
+    index_content.append("   * - Resource")
+    index_content.append("     - Description")
+    
+    for tag in sorted(tag_groups.keys()):
+        desc = tag_descs.get(tag, tag_descriptions.get(tag, ''))
+        index_content.append(f"   * - :doc:`{sanitize_filename(tag)}`")
+        index_content.append(f"     - {desc}")
+    
     index_content.append("")
     index_content.append(".. toctree::")
     index_content.append("   :maxdepth: 2")
-    index_content.append("   :caption: API Endpoints:")
+    index_content.append("   :hidden:")
     index_content.append("")
     
     # Generate documentation for each tag
@@ -293,16 +693,16 @@ def main():
             f.write(doc_content)
         
         # Add to index
-        index_content.append(f"   api/{filename}")
+        index_content.append(f"   {filename}")
         
         print(f"Generated documentation for {tag} ({len(endpoints)} endpoints)")
     
     # Save index file
-    index_path = Path(__file__).parent / "api" / "index.rst"
+    index_path = api_dir / "index.rst"
     with open(index_path, 'w') as f:
         f.write("\n".join(index_content))
     
-    print(f"\nGenerated documentation for {len(tag_groups)} tags")
+    print(f"\nGenerated enhanced documentation for {len(tag_groups)} tags")
     print(f"Documentation saved in: {api_dir}")
