+++ docs/requirements.txt
@@ -4,3 +4,4 @@ sphinx-autodoc-typehints>=1.19
 myst-parser>=0.18
 sphinx-copybutton>=0.5
 sphinx-togglebutton>=0.3
+sphinx-tabs>=3.4
